[Feature] Add DTO docs and split ProfileUpdateDTO into separate file
